+++ Library.BLL/Interfaces/IBookService.cs
@@ -4,6 +4,6 @@ namespace Library.BLL.Interfaces
 {
     public interface IBookService : IGenericService<BookModel>
     {
-        Task<BookModel> GetByISBNAsync(int isbn, CancellationToken token);
+        Task<BookModel> GetByISBNAsync(long isbn, CancellationToken token);
     }
 }
