+++ Library.DAL/Interfaces/IBookRepository.cs
@@ -4,6 +4,6 @@ namespace Library.DAL.Interfaces
 {
     public interface IBookRepository : IGenericRepository<BookEntity>
     {
-        Task<BookEntity> GetByISBNAsync(int isbn, CancellationToken token);
+        Task<BookEntity> GetByISBNAsync(long isbn, CancellationToken token);
     }
 }
