+++ Library.DAL/Repositories/BookRepository.cs
@@ -8,7 +8,7 @@ namespace Library.DAL.Repositories
     public class BookRepository : GenericRepository<BookEntity>, IBookRepository
     {
         public BookRepository(LibraryContext context) : base(context) { }
-        public Task<BookEntity> GetByISBNAsync(int isbn, CancellationToken token)
+        public Task<BookEntity> GetByISBNAsync(long isbn, CancellationToken token)
         {
             return dbSet.FirstAsync(b => b.ISBN == isbn, token);
         }
